flask_shop/IIIflask_shop/requirements.txt
alembic==1.17.2
aniso8601==10.0.1
blinker==1.9.0
click==8.3.1
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
distlib==0.3.8
filelock==3.15.4
Flask==3.1.2
Flask-Cors==3.0.10
Flask-Migrate==4.0.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
fonttools==4.53.1
greenlet==3.3.0
imutils==0.5.4
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.7
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.9.2
myutils==0.0.21
numpy==1.23.1
opencv-python==4.5.5.62
packaging==24.1
pbr==6.1.0
pillow==10.4.0
platformdirs==4.2.2
pupil-apriltags==1.0.4.post10
PyJWT==2.10.1
PyMySQL==1.0.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.16.0
SQLAlchemy==1.4.52
stevedore==5.3.0
tomli==2.3.0
typing_extensions==4.15.0
virtualenv==20.26.3
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
virtualenvwrapper-win==1.2.7
waitress==3.0.2
Werkzeug==3.1.4
